Fix Paths with Whitespace Errors
This changes how the `dbfile` string is created. Rather than using {URL}.String() in net/url (in which, paths get escaped so whitespaces are replaced with %20), this adds a function that reconstructs the path from URL.Path and appending the encoded query parameters on the end if any query parameters are present.

This is my attempt at a fix for issue #499 (which seems prematurely closed, as its still an issue).  Also includes some test cases to ensure the new function produces a the expected results based on a few different inputs(including paths with spaces, and paths with and without query arguments).