feat: add check command to check pending migrations
The check command checks for pending migrations and returns an error if
there are any.

This allows a user to use it in a script where the user doesn't want to
run the migrations but still ensure that there are none pending before
continuing.

For example, when using a Kubernetes Job to perform migrations while a
Deployment runs the service itself, a user may want to prevent the
Deployment from starting fully before the migrations have completed.